+++ src/index.ts
@@ -1,4 +1,4 @@
 export * from "./utils.ts";
 export * from "./kupid/kupid.ts";
-export * from "./type-b/info.ts";
+export * from "./type-b/index.ts";
 export * from "./type.d.ts";
