+++ src/type-b/index.ts
@@ -1,0 +1,2 @@
+export * from "./info.ts";
+export * from "./media.ts";
