+++ src/type-b/media.ts
@@ -1,5 +1,5 @@
-import { trim, convertRelativeImgPath } from "../utils.ts";
 import type { NoticeInfo } from "../type.d.ts";
+import { getNoticeContent, parseNoticeRowsFromTable } from "./type-b.ts";
 
 export type MediaCollegeType =
   | "미디어학부 공지사항"
@@ -21,11 +21,9 @@ const MEDIA_PATH_MAP: Record<MediaCollegeType, string> = {
 };
 
 const MEDIA_URL_MAP: Record<MediaCollegeType, string> = {} as Record<MediaCollegeType, string>;
-for (const type in MEDIA_PATH_MAP) {
-  MEDIA_URL_MAP[type as MediaCollegeType] = `${MEDIA_BASE_URL}${
-    MEDIA_PATH_MAP[type as MediaCollegeType]
-  }`;
-}
+for (const type in MEDIA_PATH_MAP)
+  MEDIA_URL_MAP[type as MediaCollegeType] =
+    MEDIA_BASE_URL + MEDIA_PATH_MAP[type as MediaCollegeType];
 
 /**
  *
@@ -41,77 +39,4 @@ export async function fetchMediaNotices(type: MediaCollegeType): Promise<NoticeI
   return Promise.all(
     noticeInfo.map(async (info) => ({ ...info, content: await getNoticeContent(info.url) }))
   );
-}
-
-function parseNoticeRowsFromTable(html: string, hostPath: string): Omit<NoticeInfo, "content">[] {
-  const tableRegex = /<tbody>([\s\S]+?)<\/tbody>/;
-  const tableMatch = html.match(tableRegex);
-
-  if (!tableMatch) {
-    throw new Error("Failed to parse tbody");
-  }
-  const rowRegex = /<tr[^>]*>([\s\S]+?)<\/tr>/g;
-  const rows = tableMatch[1].matchAll(rowRegex);
-  if (!rows) throw new Error("Failed to parse rows");
-
-  const result: Omit<NoticeInfo, "content">[] = [];
-
-  for (const row of rows) {
-    const td = row[1].match(/<td[^>]*>([\s\S]+?)<\/td>/g);
-    if (!td) throw new Error("Failed to parse td");
-
-    const title = trim(
-      td[1]
-        .replace(/<[^>]+>/g, "")
-        .replace(/(&nbsp;)+/g, " ")
-        .replace(/\n/g, "")
-        .replace(/\t/g, "")
-        .replace(/&amp;/g, "&")
-    ).trim();
-
-    const urlMatch = td[1].match(/href="([^"]+?)"/);
-    if (!urlMatch) throw new Error("Failed to parse url");
-    const url = `${hostPath}${urlMatch[1].replace(/&amp;/g, "&")}`;
-
-    const idMatch = url.match(/articleNo=(\d+)/);
-    if (!idMatch) throw new Error("Failed to parse id");
-    const id = idMatch[1];
-
-    const writer = td[2].replace(/<[^>]+>/g, "");
-
-    const date = td[4].replace(/<[^>]+>/g, "");
-    result.push({
-      title,
-      url,
-      id,
-      writer,
-      date,
-    });
-  }
-  return result;
-}
-
-async function getNoticeContent(url: string): Promise<string> {
-  const response = await fetch(url);
-  const html = await response.text();
-  const td = html.match(/<td[^>]*>([\s\S]+?)<\/td>/g);
-  if (!td) throw new Error("Failed to parse td");
-
-  const title = td[0].replace(/<[^>]+>/g, "").replace(/&amp;/g, "&");
-
-  const content = convertRelativeImgPath(
-    td[1]
-      .replace(/&nbsp;/g, " ")
-      .replace(/(\t)+/g, " ")
-      .replace(/(\n)+/g, "\n"),
-    MEDIA_BASE_URL
-  );
-
-  const files = td[2].replace(/&amp;/g, "&");
-  const [hostPath, _] = url.split("?");
-  const filePaths = files.replaceAll(`href="`, `href="${hostPath}`);
-  if (!filePaths) throw new Error("Failed to parse files");
-
-  const result = `<h3>${title}</h3>${content}${filePaths}`;
-  return result;
 }
